# Conflicts:
#	Web/Web.csproj

+++ Domain/SP/BE/Output/SPOutput_BE_GetOrderInfoBeforeModify.cs
@@ -9,7 +9,7 @@ namespace Domain.SP.BE.Output
 {
     public class SPOutput_BE_GetOrderInfoBeforeModify : SPOutput_Base
     {
-        public Int16 hasModify   {set;get;}
+        public Int64 hasModify   {set;get;}
         public string ModifyTime  {set;get;}
         public string ModifyUserID{set;get;}
         public string LastStartTime { set; get; }
